project/requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
ecdsa==0.19.1
fastapi==0.110.0
flask==3.1.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.12
gitpython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.27.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markupsafe==3.0.2
narwhals==1.32.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pika==1.3.2
pillow==11.1.0
protobuf==5.29.4
psycopg2==2.9.10
pyarrow==19.0.1
pyasn1==0.4.8
pycparser==2.22
pydantic==2.6.3
pydantic-core==2.16.3
pydeck==0.9.1
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-jose==3.4.0
python-multipart==0.0.9
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
rsa==4.9
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.36.3
streamlit==1.44.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
typing-extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.27.1
werkzeug==3.1.3
